Support dynamic page loading
If there's a change in the DOM, the last AJAX request is aborted and a new process request is sent, in order to process pages that are dynamically loaded.
The observer must be disconnected before any DOM changes in the content script and enabled immediately after again.
CSS rules were also changed as the tooltip now has class "health-translator" in the outer wrapper, so it's not scoped.