Change replace mode in order to change only text nodes
The language is first detected on the client and the text nodes are sent to the server, instead of sending and replacing the whole <body>. The previous method has problems and disadvantages: DOM references are lost, event handlers are lost and non-serialisable information is lost (http://stackoverflow.com/questions/4060056/jquery-find-replace-without-changing-original-text/4060635#4060635). However, the new method implies that many more requests are sent to the server.