Change the script injection approach
Scripts are now injected all the time and when the DOM is loaded it automatically executes if on automatic mode. If manual mode and the user clicks the icon before the DOM loaded, a flag is set up and verified after the load.
Also, injection of Bootstrap is avoided if already present in the host page.